mejoras, data_type general, int

+++ src/list/ft_mtcollect_type.c
@@ -6,7 +6,7 @@
 /*   By: danpalac <danpalac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/27 18:58:49 by danpalac          #+#    #+#             */
-/*   Updated: 2024/12/02 11:51:48 by danpalac         ###   ########.fr       */
+/*   Updated: 2024/12/10 12:27:03 by danpalac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ t_mt	*ft_mtcollect_node_type_mt(t_mt *mt, t_node_type type)
 		return (NULL);
 	collected = NULL;
 	sub = NULL;
-	if (mt->values.node_type == type)
+	if (mt->values.node_type == (int)type)
 	{
 		sub = ft_mtdup(mt);
 		ft_mtpush_back(&sub, &collected);
@@ -55,7 +55,7 @@ t_mt	*ft_mtcollect_data_type_mt(t_mt *mt, t_data_type type)
 		return (NULL);
 	collected = NULL;
 	sub = NULL;
-	if (mt->values.data_type == type)
+	if (mt->values.data_type == (int)type)
 	{
 		sub = ft_mtdup(mt);
 		ft_mtpush_back(&sub, &collected);
